requirements.txt
streamlit==1.46.1
langchain==0.3.26
langchain-openai==0.3.27
langchain-community==0.3.26
faiss-cpu==1.11.0
python-dotenv==1.1.1
PyMuPDF==1.26.1
openai==1.92.3
pypdf==5.6.1
